- deletes large file
- WIP 2
I see, they use noisy linear layers for exploration.
try to increase --noisy-std 0.5, 1, 2, 5, 10... until you find that the agent can reach the flag with a similar probability as if you would apply random actions.
It that will not help, you can extend that main.py with a random action selection for 90% of actions. Then decrease this % linearly until 0% at the end of training.

- I see, they use noisy linear layers for exploration.
try to increase --noisy-std 0.5, 1, 2, 5, 10... until you find that the agent can reach the flag with a similar probability as if you would apply random actions.
It that will not help, you can extend that main.py with a random action selection for 90% of actions. Then decrease this % linearly until 0% at the end of training.

- Found frameskip number for random action, but in Rainbow it's not working

- Manual frameskip in progress

Frameskip : the problem is not in learning. The problem is that the agent can not get reward with random actions withing 200 steps. The agent should be able to reach the flag with random actions at least every 5-10th episode. For that you can adjust the frameskip value. Find aframeskip value (e.g. 5, 10, 20...) such that it allows you (keyboard agent) to solve the task within a sequence of 4-5 keyboard actions (make rendering pause before each keyboard action). Give me a SlackCall when you find the right frameskip value (do not run learning so far)